refactor: split components into files
