+++ src/components/objectSearch/results/grids/index.tsx
@@ -1,0 +1,1 @@
+export * from "./grids";
